+++ src/fitplusbike.cpp
@@ -538,7 +538,7 @@ void fitplusbike::characteristicChanged(const QLowEnergyCharacteristic &characte
 #endif
         {
             if (Flags.heartRate && !disable_hr_frommachinery && newValue.length() > index) {
-                Heart = ((double)((newValue.at(index))));
+                Heart = ((double)(((uint8_t)newValue.at(index))));
                 // index += 1; // NOTE: clang-analyzer-deadcode.DeadStores
                 qDebug() << (QStringLiteral("Current Heart: ") + QString::number(Heart.value()));
             } else {
