+++ src/solef80treadmill.cpp
@@ -674,7 +674,7 @@ void solef80treadmill::characteristicChanged(const QLowEnergyCharacteristic &cha
             if (Flags.heartRate) {
                 if (index < newValue.length()) {
 
-                    heart = ((double)((newValue.at(index))));
+                    heart = ((double)(((uint8_t)newValue.at(index))));
                     emit debug(QStringLiteral("Current Heart: ") + QString::number(heart));
                 } else {
                     emit debug(QStringLiteral("Error on parsing heart!"));
