+++ src/technogymmyruntreadmill.cpp
@@ -395,7 +395,7 @@ void technogymmyruntreadmill::characteristicChanged(const QLowEnergyCharacterist
             if (Flags.heartRate) {
                 if (index < newValue.length()) {
 
-                    heart = ((double)((newValue.at(index))));
+                    heart = ((double)(((uint8_t)newValue.at(index))));
                     emit debug(QStringLiteral("Current Heart: ") + QString::number(heart));
                 } else {
                     emit debug(QStringLiteral("Error on parsing heart!"));
